Factor out `_aggregate_proportions` as the shared sample-level primitive
The "weighted sum of per-observation proportion rows per group" was implemented
twice -- once in `GPCCA._obsm_composition` (plotting) and once in
`_map_names_and_colors_from_proportions` (naming), transposed and with slightly
different syntax. Extract it into a single reusable, explicitly-documented
`_aggregate_proportions(proportions, groups, weights)` helper so the obsm/sample-
level code path has one source of truth that future features can build on.

Layering after this change:
- `_utils.py` owns the obsm/sample-level toolkit: `_resolve_composition_key`,
  `_obsm_proportions`, `_check_proportions`, `_obsm_proportion_weights`,
  `_aggregate_proportions`, and the proportions front-end
  `_map_names_and_colors_from_proportions` (which composes `_aggregate_proportions`
  with the colors tail).
- `_colors.py` stays purely about color/naming: `_names_and_colors_from_association`
  (reusable association-table -> unique names+colors) and the hard-count
  `_map_names_and_colors`. It no longer knows about proportions.
- `GPCCA._obsm_composition` and macrostate naming both consume
  `_aggregate_proportions`, so they share identical semantics.

Moves the proportions unit tests to `tests/test_utils.py` and adds a dedicated
`TestAggregateProportions` suite.
